SQL Analysis of Website Sessions and Orders
NOTE REGARDING SQL PROJECT

To whom it may concern,

Please be advised that in the SQL project presented, specific names and dates have been modified to ensure the confidentiality and privacy of our client. This decision was made in compliance with the Non-Disclosure Agreement (NDA) that I am bound to. Any resemblance to real names or dates is purely coincidental.

These modifications will not affect the functionality or accuracy of the project but are intended purely for the purpose of upholding our commitment to client privacy. Should you have any questions or concerns regarding this, please feel free to reach out to me directly.

Thank you for your understanding.